server/requirements.txt
google-genai>=1.0.0
websockets>=12.0
python-dotenv>=1.0.0
